+++ core/Piwik.php
@@ -427,7 +427,10 @@ class Piwik
      * Helper method user to set the current as superuser.
      * This should be used with great care as this gives the user all permissions.
      *
+     * This method is deprecated, use {@link Access::doAsSuperUser()} instead.
+     *
      * @param bool $bool true to set current user as Super User
+     * @deprecated
      */
     public static function setUserHasSuperUserAccess($bool = true)
     {