+++ plugins/Installation/Controller.php
@@ -196,14 +196,16 @@ class Controller extends \Piwik\Plugin\ControllerAdmin
             $view->tablesInstalled     = implode(', ', $tablesInstalled);
             $view->someTablesInstalled = true;
 
-            Access::getInstance();
-            Piwik::setUserHasSuperUserAccess();
-            if ($this->hasEnoughTablesToReuseDb($tablesInstalled) &&
-                count(APISitesManager::getInstance()->getAllSitesId()) > 0 &&
-                count(APIUsersManager::getInstance()->getUsers()) > 0
-            ) {
-                $view->showReuseExistingTables = true;
-            }
+            $self = $this;
+            Access::doAsSuperUser(function () use ($self, $tablesInstalled, $view) {
+                Access::getInstance();
+                if ($self->hasEnoughTablesToReuseDb($tablesInstalled) &&
+                    count(APISitesManager::getInstance()->getAllSitesId()) > 0 &&
+                    count(APIUsersManager::getInstance()->getUsers()) > 0
+                ) {
+                    $view->showReuseExistingTables = true;
+                }
+            });
         } else {
 
             DbHelper::createTables();
@@ -258,8 +260,11 @@ class Controller extends \Piwik\Plugin\ControllerAdmin
     {
         $this->checkPiwikIsNotInstalled();
 
-        $this->initObjectsToCallAPI();
-        if (count(APIUsersManager::getInstance()->getUsersHavingSuperUserAccess()) > 0) {
+        $superUserAlreadyExists = Access::doAsSuperUser(function () {
+            return count(APIUsersManager::getInstance()->getUsersHavingSuperUserAccess()) > 0;
+        });
+
+        if ($superUserAlreadyExists) {
             $this->redirectToNextStep('setupSuperUser');
         }
 
@@ -301,9 +306,11 @@ class Controller extends \Piwik\Plugin\ControllerAdmin
     {
         $this->checkPiwikIsNotInstalled();
 
-        $this->initObjectsToCallAPI();
+        $siteIdsCount = Access::doAsSuperUser(function () {
+            return count(APISitesManager::getInstance()->getAllSitesId());
+        });
 
-        if (count(APISitesManager::getInstance()->getAllSitesId()) > 0) {
+        if ($siteIdsCount > 0) {
             // if there is a already a website, skip this step and trackingCode step
             $this->redirectToNextStep('trackingCode');
         }
@@ -322,7 +329,10 @@ class Controller extends \Piwik\Plugin\ControllerAdmin
             $ecommerce = (int)$form->getSubmitValue('ecommerce');
 
             try {
-                $result = APISitesManager::getInstance()->addSite($name, $url, $ecommerce);
+                $result = Access::doAsSuperUser(function () use ($name, $url, $ecommerce) {
+                    return APISitesManager::getInstance()->addSite($name, $url, $ecommerce);
+                });
+
                 $params = array(
                     'site_idSite' => $result,
                     'site_name' => urlencode($name)
@@ -461,14 +471,6 @@ class Controller extends \Piwik\Plugin\ControllerAdmin
     private function getParam($name)
     {
         return Common::getRequestVar($name, false, 'string');
-    }
-
-    /**
-     * Instantiate access and log objects
-     */
-    private function initObjectsToCallAPI()
-    {
-        Piwik::setUserHasSuperUserAccess();
     }
 
     /**
@@ -633,13 +635,12 @@ class Controller extends \Piwik\Plugin\ControllerAdmin
 
     private function createSuperUser($login, $password, $email)
     {
-        $this->initObjectsToCallAPI();
-
-        $api = APIUsersManager::getInstance();
-        $api->addUser($login, $password, $email);
-
-        $this->initObjectsToCallAPI();
-        $api->setSuperUserAccess($login, true);
+        $self = $this;
+        Access::doAsSuperUser(function () use ($self, $login, $password, $email) {
+            $api = APIUsersManager::getInstance();
+            $api->addUser($login, $password, $email);
+            $api->setSuperUserAccess($login, true);
+        });
     }
 
     private function hasEnoughTablesToReuseDb($tablesInstalled)
@@ -702,16 +703,16 @@ class Controller extends \Piwik\Plugin\ControllerAdmin
     protected function updateComponents()
     {
         Access::getInstance();
-        Piwik::setUserHasSuperUserAccess();
 
-        $updater = new Updater();
-        $componentsWithUpdateFile = CoreUpdater::getComponentUpdates($updater);
+        return Access::doAsSuperUser(function () {
+            $updater = new Updater();
+            $componentsWithUpdateFile = CoreUpdater::getComponentUpdates($updater);
 
-        if (empty($componentsWithUpdateFile)) {
-            return false;
-        }
-        $result = CoreUpdater::updateComponents($updater, $componentsWithUpdateFile);
-        return $result;
+            if (empty($componentsWithUpdateFile)) {
+                return false;
+            }
+            $result = CoreUpdater::updateComponents($updater, $componentsWithUpdateFile);
+            return $result;
+        });
     }
-
 }