+++ tests/PHPUnit/Fixtures/SqlDump.php
@@ -7,6 +7,7 @@
  */
 namespace Piwik\Tests\Fixtures;
 
+use Piwik\Access;
 use Piwik\ArchiveProcessor\Rules;
 use Piwik\Config;
 use Piwik\Db;
@@ -79,7 +80,7 @@ class SqlDump extends Fixture
         Rules::setBrowserTriggerArchiving(true);
 
         // reload access
-        Piwik::setUserHasSuperUserAccess();
+        Access::getInstance()->reloadAccess();
 
         $this->getTestEnvironment()->configOverride = array(
             'database' => array(